Folder Restructure + Re-gitignore
Decided to move all of the app/game files to the repo root, I think this is a little more easier to work with. Usually you would find the package.json for the project at the root as well.

Also, I updated Electron to the latest version. It was actually the only package for the project that had a newer update available.