On the Way to the Web!
Removed Electron and extra sound playing dependencies. I want to try making the original game work on the web exclusively, so I will just be keeping around Howler for now. I think it seems to be an audio playing library, so I may be able to reimplement it with simple vanilla JS for the frontend.