Howler Dependency Change
Found a CDN hosting source for the Howler sound library! Now this means it doesn't have to be included and updated alongside the game. Because of this, there aren't any more large usecases to have the npm package files, so I removed those. They only had metadata in them at this point because there weren't anymore npm packages to keep track of.

Now that everything for the game is in one place, it can be played exclusively on the web without any issues! Off to GitHub Pages we go...